tools/pyproject.toml
[project]
name = "plurk-backup-tools"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "playwright>=1.40.0",
]

[project.scripts]
plurk-tools = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
